separated tests into different files for each endpoint
